+++ core/src/clipboard/mac.rs
@@ -1,6 +1,6 @@
 use anyhow::Result;
 use secrecy::SecretString;
 
-pub(crate) fn copy_to_clip_board(p0: SecretString, p1: Option<usize>) -> Result<()> {
+pub(crate) fn copy_to_clip_board(mut secret: SecretString, timeout: Option<usize>) -> Result<()> {
     todo!()
 }